use the modelapi objects member to get objects
This allows for pre-filtering of objects served through the api.
Note however that it doesn't filter objects which are returned through
api_method or by following foreign keys!
Also removed the model from the registered_types, in preparation for a
fuller decoupling of an api type from django models. The django model is
still used for discovering the api_methods however.